+++ .github/workflows/test.yml
@@ -28,3 +28,24 @@ jobs:
         uses: codecov/codecov-action@v3
         env:
           CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+
+  integration-test:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Set up Go
+        uses: actions/setup-go@v2
+        with:
+          go-version: 1.21
+
+      - name: Verify dependencies
+        run: go mod verify
+
+      - name: Run inline tests
+        run: |
+          go test --tags=inline ./pkg/test/...
+
+      - name: Run external tests
+        run: |
+          go test --tags=external ./pkg/test/...
