+++ pkg/test/delete_test.go
